Update 11
The only changes i made before the presentation were changing the background colour of the gui to all be green as they were different colours for testing, I also added back in the addPlayer method on line 148 of the BlackjackGUI as it had been taken out while i tried to get the cards to change.

After the presentation i commented out where i called the GUI in BlackJackDriver on line 40 as it was not supposed to be in there.